HashDicts

Like Base.Dict, but slower. This package defines a single AbstractDict type which relies on key hashnig and equality comparison.

HashDicts.HashDict — Type

HashDict{K, V} <: AbstractDict{K, V}

Dict-like behavior, stores key-value pairs by key hash.

Resizing and managing capacity

Base.resize! — Method

Base.resize!(d::HashDict)

Double the size of the data underlying the hash dict. Called when inserting an item goes above LOAD_FACTOR.

Base.resize! — Method

Base.resize!(d::HashDict{K, V}, capacity::Integer)

Resize a HashDict by adding a given capacity. Becomes a no-op if capacity <= 0.

HashDicts.LOAD_FACTOR — Constant

Maximum average number of items per bucket. Resizing is performed when this average is reached.